July 2021 update

+++ skyalmanac.py
@@ -31,16 +31,31 @@ import increments
 from alma_skyfield import init_sf
 
 
-def makePDF(fn, msg = ""):
-    command = 'pdflatex {}'.format(fn + ".tex")
-    os.system(command)
-    print("finished" + msg)
+def makePDF(args, fn, msg = ""):
+    command = 'pdflatex {}'.format(args + fn + ".tex")
+    if args == "":
+        os.system(command)
+        print("finished" + msg)
+    else:
+        returned_value = os.system(command)
+        if returned_value != 0:
+            if msg != "":
+                print("ERROR detected while" + msg)
+            else:
+                print("!!   ERROR detected while creating PDF file   !!")
+                print("!! Append '-v' or '-log' for more information !!")
+        else:
+            if msg != "":
+                print("finished" + msg)
+            else:
+                print("finished creating PDF")
     return
 
 def tidy_up(fn):
-    os.remove(fn + ".tex")
-    if os.path.isfile(fn + ".log"):
-        os.remove(fn + ".log")
+    if not keeptex: os.remove(fn + ".tex")
+    if not keeplog:
+        if os.path.isfile(fn + ".log"):
+            os.remove(fn + ".log")
     if os.path.isfile(fn + ".aux"):
         os.remove(fn + ".aux")
     return
@@ -49,6 +64,24 @@ def tidy_up(fn):
 ##Main##
 if sys.version_info[0] < 3:
     raise Exception("Must be using Python 3")
+
+try:
+    arg = sys.argv[1]
+except IndexError:
+    arg = ""
+if len(sys.argv) > 2 or not (arg == "" or arg == "-v" or arg == "-log" or arg == "-tex"):
+    print("One optional command line parameter is permitted:")
+    print(" python -m skyalmanac -v")
+    print(" ... to send pdfTeX output to the terminal")
+    print(" python -m skyalmanac -log")
+    print(" ... to keep the log file")
+    print(" python -m skyalmanac -tex")
+    print(" ... to keep the tex file")
+    sys.exit(0)
+
+args = "" if arg == "-v" else "-interaction=batchmode -halt-on-error "
+keeplog = True if arg == "-log" else False
+keeptex = True if arg == "-tex" else False
 
 d = datetime.datetime.utcnow().date()
 first_day = datetime.date(d.year, d.month, d.day)
@@ -64,7 +97,7 @@ if config.dockerized:
     config.pgsz = os.getenv('PGSZ', config.pgsz)
     config.moonimg = os.getenv('MOONIMG', str(config.moonimg))
     config.ephndx = os.getenv('EPHNDX', str(config.ephndx))
-    if config.ephndx not in set(['0', '1', '2']):
+    if config.ephndx not in set(['0', '1', '2', '3', '4']):
         ephERR = True
     else:
         config.ephndx = int(config.ephndx)
@@ -92,7 +125,7 @@ else:
     #first_day = datetime.date(2023, 6, 24)	# for testing a specific date
     #d = first_day							# for testing a specific date
     spad = "./"
-    if config.ephndx not in set([0, 1, 2]):
+    if config.ephndx not in set([0, 1, 2, 3, 4]):
         ephERR = True
     config.moonimg = str(config.moonimg)
     config.useIERS = str(config.useIERS)
@@ -207,7 +240,8 @@ if s in set(['1', '2', '3', '4', '5']):
             print(msg)
 ##            config.writeLOG(msg)
             first_day = datetime.date(yearint, 1, 1)
-            fn = "almanac{}{}".format(ff,year+DecFmt)
+            ff = "tradna_" if config.tbls != 'm' else "modna_"
+            fn = "{}{}".format(ff,year+DecFmt)
             outfile = open(f_prefix + fn + ".tex", mode="w", encoding="utf8")
             outfile.write(tables.almanac(first_day,122))
             outfile.close()
@@ -217,7 +251,7 @@ if s in set(['1', '2', '3', '4', '5']):
 ##            config.writeLOG("\n\n" + msg + "\n")
             print()
             if config.dockerized: os.chdir(os.getcwd() + f_postfix)     # DOCKER ONLY
-            makePDF(fn, " creating nautical almanac for {}".format(year))
+            makePDF(args, fn, " creating nautical almanac for {}".format(year))
             tidy_up(fn)
             if config.dockerized: os.chdir(docker_main)     # reset working folder to code folder
 ##        config.closeLOG()     # close log after the for-loop
@@ -228,12 +262,13 @@ if s in set(['1', '2', '3', '4', '5']):
             msg = "\nCreating the sun tables for the year {}\n".format(year)
             print(msg)
             first_day = datetime.date(yearint, 1, 1)
-            fn = "sunalmanac{}{}".format(ff,year+DecFmt)
+            ff = "tradst_" if config.tbls != 'm' else "modst_"
+            fn = "{}{}".format(ff,year+DecFmt)
             outfile = open(f_prefix + fn + ".tex", mode="w", encoding="utf8")
             outfile.write(suntables.almanac(first_day,25))
             outfile.close()
             if config.dockerized: os.chdir(os.getcwd() + f_postfix)     # DOCKER ONLY
-            makePDF(fn, " creating sun tables for {}".format(year))
+            makePDF(args, fn, " creating sun tables for {}".format(year))
             tidy_up(fn)
             if config.dockerized: os.chdir(docker_main)     # reset working folder to code folder
 
@@ -242,7 +277,8 @@ if s in set(['1', '2', '3', '4', '5']):
         start = time.time()
         msg = "\nCreating nautical almanac tables - from {}\n".format(sdmy)
         print(msg)
-        fn = "almanac{}{}".format(ff,symd+DecFmt)
+        ff = "tradna_" if config.tbls != 'm' else "modna_"
+        fn = "{}{}".format(ff,symd+DecFmt)
         outfile = open(f_prefix + fn + ".tex", mode="w", encoding="utf8")
         outfile.write(tables.almanac(first_day,2))
         outfile.close()
@@ -253,18 +289,19 @@ if s in set(['1', '2', '3', '4', '5']):
 ##        config.closeLOG()
         print()
         if config.dockerized: os.chdir(os.getcwd() + f_postfix)     # DOCKER ONLY
-        makePDF(fn)
+        makePDF(args, fn)
         tidy_up(fn)
 
     elif s == '4':      # Sun tables only    - 30 days from today
         msg = "\nCreating the sun tables - from {}\n".format(sdmy)
         print(msg)
-        fn = "sunalmanac{}{}".format(ff,symd+DecFmt)
+        ff = "tradst_" if config.tbls != 'm' else "modst_"
+        fn = "{}{}".format(ff,symd+DecFmt)
         outfile = open(f_prefix + fn + ".tex", mode="w", encoding="utf8")
         outfile.write(suntables.almanac(first_day,2))
         outfile.close()
         if config.dockerized: os.chdir(os.getcwd() + f_postfix)     # DOCKER ONLY
-        makePDF(fn)
+        makePDF(args, fn)
         tidy_up(fn)
 
     elif s == '5':
@@ -275,7 +312,7 @@ if s in set(['1', '2', '3', '4', '5']):
         outfile.write(increments.makelatex())
         outfile.close()
         if config.dockerized: os.chdir(os.getcwd() + f_postfix)     # DOCKER ONLY
-        makePDF(fn)
+        makePDF(args, fn)
         tidy_up(fn)
 
 else:
